requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
cloudinary==1.30.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==1.0.0
Django==4.1.2
django-autoslug==1.9.8
django-ckeditor==6.5.1
django-cloudinary-storage==0.3.0
django-debug-toolbar==3.7.0
django-import-export==3.0.1
django-jazzmin==2.5.0
django-js-asset==2.0.0
django-social-share==2.3.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
psycopg2-binary==2.9.4
python-decouple==3.6
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
sweetify==2.3.1
tablib==3.2.1
urllib3==1.26.12
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
